Move isNullOrTemporalAmount
